Ver 1.1 Fixed multiple issues can solve the first 2 boards.
The main issue is the logic surrounding Action class.
a bit confusing as the code was designed to center around the empty tile. But, we have to return the operation used on the non-empty tile so there are issues following the logic there.
Still not satisfied with the amount of gets and sets. Seems like it bode  well with the OOP paradigm.

Things to do:
Fix action class - moved tile / empty tile are confusing.
Think of re-arranging the data.
Add heuristic value (only one that comes to mind is the suggested misplaced tiles).